Ensure to always use field types, based on widget streams in dashboard widgets.
