+++ graylog2-web-interface/src/views/components/widgets/Widget.aggregations.test.tsx
@@ -149,7 +149,6 @@ describe('Aggregation Widget', () => {
           <WidgetContext.Provider value={propsWidget}>
             <Widget widget={propsWidget}
                     id="widgetId"
-                    fields={Immutable.List([])}
                     onPositionsChange={() => {}}
                     title="Widget Title"
                     position={new WidgetPosition(1, 1, 1, 1)}
