+++ graylog2-web-interface/src/views/components/widgets/Widget.tsx
@@ -16,10 +16,10 @@
  */
 import * as React from 'react';
 import { useCallback, useContext, useMemo, useState } from 'react';
-import type * as Immutable from 'immutable';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
+import useActiveQueryId from 'views/hooks/useActiveQueryId';
 import type { BackendWidgetPosition, WidgetResults } from 'views/types';
 import { useStore } from 'stores/connect';
 import { widgetDefinition } from 'views/logic/Widgets';
@@ -28,7 +28,6 @@ import { WidgetActions } from 'views/stores/WidgetStore';
 import { TitlesActions } from 'views/stores/TitlesStore';
 import { ViewStore } from 'views/stores/ViewStore';
 import { RefreshActions } from 'views/stores/RefreshStore';
-import type FieldTypeMapping from 'views/logic/fieldtypes/FieldTypeMapping';
 import WidgetModel from 'views/logic/widgets/Widget';
 import type WidgetPosition from 'views/logic/widgets/WidgetPosition';
 import type { Rows } from 'views/logic/searchtypes/pivot/PivotHandler';
@@ -39,6 +38,7 @@ import IfDashboard from 'views/components/dashboard/IfDashboard';
 import type WidgetConfig from 'views/logic/widgets/WidgetConfig';
 import type { FieldTypeMappingsList } from 'views/logic/fieldtypes/types';
 import useWidgetResults from 'views/components/useWidgetResults';
+import FieldTypesContext from 'views/components/contexts/FieldTypesContext';
 
 import WidgetFrame from './WidgetFrame';
 import WidgetHeader from './WidgetHeader';
@@ -55,7 +55,6 @@ export type Props = {
   id: string,
   widget: WidgetModel,
   editing: boolean,
-  fields: Immutable.List<FieldTypeMapping>,
   title: string,
   position: WidgetPosition,
   onPositionsChange: (position: BackendWidgetPosition) => void,
@@ -79,17 +78,25 @@ const _hasOwnEditSubmitButton = (type) => {
   return widgetDefinition(type).hasEditSubmitButton;
 };
 
+const useQueryFieldTypes = () => {
+  const fieldTypes = useContext(FieldTypesContext);
+  const queryId = useActiveQueryId();
+
+  return useMemo(() => fieldTypes.queryFields.get(queryId, fieldTypes.all), [fieldTypes.all, fieldTypes.queryFields, queryId]);
+};
+
 const WidgetFooter = styled.div`
   width: 100%;
   display: flex;
   justify-content: flex-end;
 `;
 
-type VisualizationProps = Pick<Props, 'title' | 'id' | 'widget' | 'fields' | 'editing'> & {
+type VisualizationProps = Pick<Props, 'title' | 'id' | 'widget' | 'editing'> & {
   queryId: string,
   setLoadingState: (loading: boolean) => void,
   onToggleEdit: () => void,
   onWidgetConfigChange: (newWidgetConfig: WidgetConfig) => Promise<Widgets>,
+  fields: FieldTypeMappingsList,
 };
 
 const Visualization = ({ title, id, widget, fields, queryId, editing, setLoadingState, onToggleEdit, onWidgetConfigChange }: VisualizationProps) => {
@@ -154,7 +161,8 @@ const EditWrapper = ({ children, config, editing, fields, id, onToggleEdit, onCa
   ) : children;
 };
 
-const Widget = ({ id, editing, widget, fields, title, position, onPositionsChange }: Props) => {
+const Widget = ({ id, editing, widget, title, position, onPositionsChange }: Props) => {
+  const fields = useQueryFieldTypes();
   const [loading, setLoading] = useState(false);
   const [oldWidget, setOldWidget] = useState(editing ? widget : undefined);
   const { focusedWidget, setWidgetEditing, unsetWidgetEditing } = useContext(WidgetFocusContext);
@@ -233,7 +241,6 @@ const Widget = ({ id, editing, widget, fields, title, position, onPositionsChang
 
 Widget.propTypes = {
   editing: PropTypes.bool,
-  fields: PropTypes.any.isRequired,
   id: PropTypes.string.isRequired,
   onPositionsChange: PropTypes.func.isRequired,
   title: PropTypes.string.isRequired,
